Move console realization, add config_path, create function init app
