+++ bin/console_application/main.py
@@ -1,7 +1,8 @@
-import app.bootstrap.initialization_app
+from app.console.users.users import register_command
+from app.bootstrap.initialization_app import initialize_app
 
-from app.api.console.users.users import register_command
 
 if __name__ == "__main__":
+    initialize_app(config_path="../../config.yml")
     cli = register_command()
     cli()
