+++ src/app/bootstrap/initialization_app.py
@@ -4,12 +4,14 @@ from app.factories.repository_factory import StorageRepositoryFactory, MemoryRep
 from app.repository.base import BaseUserRepository
 from app.utils.ioc import ioc
 
-config_factory_storage = ConfigFactoryStorage()
-config_factory_storage.register_factory(YmlConfigFactory())
-config = config_factory_storage.get_instance("../../config.yml")
 
-repository_factory_storage = StorageRepositoryFactory()
-repository_factory_storage.register_factory(MemoryRepositoryFactory())
-repository_factory_storage.register_factory(PostgresRepositoryFactory())
-repository = repository_factory_storage.get_instance(config.type, settings=config.settings)
-ioc.set(BaseUserRepository, repository)
+def initialize_app(config_path: str):
+    config_factory_storage = ConfigFactoryStorage()
+    config_factory_storage.register_factory(YmlConfigFactory())
+    config = config_factory_storage.get_instance(config_path)
+
+    repository_factory_storage = StorageRepositoryFactory()
+    repository_factory_storage.register_factory(MemoryRepositoryFactory())
+    repository_factory_storage.register_factory(PostgresRepositoryFactory())
+    repository = repository_factory_storage.get_instance(config.type, settings=config.settings)
+    ioc.set(BaseUserRepository, repository)
