+++ src/app/console/users/delete.py
